pipeline/requirements.txt
# F1 OmniSense — Python backend dependencies
# torch/torchvision are installed separately with --index-url for CPU-only builds

# Web framework
fastapi==0.128.4
uvicorn==0.40.0
starlette==0.52.1
python-multipart==0.0.22
anyio==4.12.1

# LLM & AI
groq==0.37.1
sentence-transformers==5.2.2
transformers==5.1.0
tokenizers==0.22.2
safetensors==0.7.0
open-clip-torch==3.2.0
einops

# Database
pymongo==4.16.0

# Data processing
numpy==2.3.5
scipy==1.17.0
scikit-learn==1.8.0
pillow==11.3.0
pydantic==2.12.5

# Document parsing
PyMuPDF==1.26.7
python-docx==1.2.0
pdfplumber==0.11.9
pypdf==5.6.0

# 3D Model Generation (Hunyuan3D via HuggingFace Spaces)
gradio-client==1.10.1

# Utilities
python-dotenv==1.2.1
requests==2.32.5
httpx==0.28.1
httpcore==1.0.9
tqdm==4.67.3

# LangChain ecosystem (pinned to local working versions)
langchain-core==1.2.16
langchain-text-splitters==1.1.0
langchain-community==0.4.1
langchain-ollama==1.0.1
langchain-groq==1.1.2
langchain-openai==1.1.10
langchain-anthropic==1.3.4

# Computer Vision (OmniVis — GPU builds only, ignored on CPU)
ultralytics>=8.3.0
opencv-python-headless>=4.9.0
supervision>=0.22.0

# OmniSuite integration
camelot-py[cv]>=1.1.0
umap-learn==0.5.11
lightgbm==4.6.0
shap==0.50.0
